split mod stream into two files
