+++ src/common/stream_async.rs
@@ -19,36 +19,10 @@
  * under the License.
  */
 
-pub(crate) mod address;
-mod credential;
-pub mod error;
-mod id;
-pub(crate) mod info;
-mod options;
-#[cfg_attr(not(feature = "sync"), path = "stream_async.rs")]
-#[cfg_attr(feature = "sync", path = "stream_sync.rs")]
-pub mod stream;
+pub use futures::{stream::BoxStream, Stream};
 
-pub(crate) use self::stream::box_stream;
-pub use self::{credential::Credential, error::Error, options::Options};
-
-pub(crate) type StdResult<T, E> = std::result::Result<T, E>;
-pub type Result<T = ()> = StdResult<T, Error>;
-
-pub(crate) type IID = id::ID;
-pub(crate) type RequestID = id::ID;
-pub(crate) type SessionID = id::ID;
-
-#[repr(C)]
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub enum SessionType {
-    Data = 0,
-    Schema = 1,
+pub(crate) fn box_stream<'a, T>(stream: impl Stream<Item = T> + Send + 'a) -> BoxStream<'a, T> {
+    Box::pin(stream) as futures::stream::BoxStream<_>
 }
 
-#[repr(C)]
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub enum TransactionType {
-    Read = 0,
-    Write = 1,
-}
+pub use tokio_stream::wrappers::UnboundedReceiverStream as NetworkStream;
