+++ src/common/stream_sync.rs
@@ -19,36 +19,29 @@
  * under the License.
  */
 
-pub(crate) mod address;
-mod credential;
-pub mod error;
-mod id;
-pub(crate) mod info;
-mod options;
-#[cfg_attr(not(feature = "sync"), path = "stream_async.rs")]
-#[cfg_attr(feature = "sync", path = "stream_sync.rs")]
-pub mod stream;
+pub use std::iter::Iterator as Stream;
 
-pub(crate) use self::stream::box_stream;
-pub use self::{credential::Credential, error::Error, options::Options};
+use tokio::sync::mpsc::UnboundedReceiver;
 
-pub(crate) type StdResult<T, E> = std::result::Result<T, E>;
-pub type Result<T = ()> = StdResult<T, Error>;
+pub type BoxStream<'a, T> = Box<dyn Stream<Item = T> + Send + 'a>;
 
-pub(crate) type IID = id::ID;
-pub(crate) type RequestID = id::ID;
-pub(crate) type SessionID = id::ID;
-
-#[repr(C)]
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub enum SessionType {
-    Data = 0,
-    Schema = 1,
+pub(crate) fn box_stream<'a, T>(stream: impl Iterator<Item = T> + Send + 'a) -> BoxStream<'a, T> {
+    Box::new(stream) as BoxStream<'a, T>
 }
 
-#[repr(C)]
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub enum TransactionType {
-    Read = 0,
-    Write = 1,
+pub struct NetworkStream<T> {
+    receiver: UnboundedReceiver<T>,
+}
+
+impl<T> NetworkStream<T> {
+    pub fn new(receiver: UnboundedReceiver<T>) -> Self {
+        Self { receiver }
+    }
+}
+
+impl<T> Iterator for NetworkStream<T> {
+    type Item = T;
+    fn next(&mut self) -> Option<Self::Item> {
+        self.receiver.blocking_recv()
+    }
 }
